Updated android `gradle` and Refactored `files`
