+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/interfaces/PermissionManagerInterface.kt
@@ -2,8 +2,8 @@ package com.lucasjosino.on_audio_query.interfaces
 
 import android.content.Context
 
-/** OnPermissionManagerInterface */
-interface OnPermissionManagerInterface {
+/** PermissionManagerInterface */
+interface PermissionManagerInterface {
     fun onPermissionStatus(context: Context? = null) : Boolean
     fun onRequestPermission()
     fun onRetryRequestPermission()
