+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/query/SongsQuery.kt
@@ -7,7 +7,7 @@ import android.net.Uri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.lucasjosino.on_audio_query.OnAudioQueryPlugin
-import com.lucasjosino.on_audio_query.query.helper.OnAudioHelper
+import com.lucasjosino.on_audio_query.query.helper.QueryHelper
 import com.lucasjosino.on_audio_query.types.checkAudiosUriType
 import com.lucasjosino.on_audio_query.types.sorttypes.checkSongSortType
 import com.lucasjosino.on_audio_query.utils.songProjection
@@ -17,11 +17,11 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
-/** OnAudiosQuery */
-class OnAudiosQuery : ViewModel() {
+/** SongsQuery */
+class SongsQuery : ViewModel() {
 
     // Main parameters
-    private val helper = OnAudioHelper()
+    private val helper = QueryHelper()
     private var selection: String? = null
 
     // None of this methods can be null.
